Various changes for testing

+++ Battle City/Point2D.cpp
@@ -10,21 +10,21 @@
 
 Point2D::Point2D() : Point2D(0, 0) {}
 
-Point2D::Point2D(int x, int y) : x(x), y(y) {}
+Point2D::Point2D(double x, double y) : x(x), y(y) {}
 
-int Point2D::getX() const {
+double Point2D::getX() const {
     return x;
 }
 
-void Point2D::setX(int x) {
+void Point2D::setX(double x) {
     this->x = x;
 }
 
-int Point2D::getY() const {
+double Point2D::getY() const {
     return y;
 }
 
-void Point2D::setY(int y) {
+void Point2D::setY(double y) {
     this->y = y;
 }
